fix build issues arising due to not casting when needed

+++ src/maestro.rs
@@ -251,7 +251,7 @@ impl Maestro {
         return write_result
             .and_then(|()| self.read(RESPONSE_SIZE))
             .and_then(move |bytes_read| {
-                return if bytes_read == RESPONSE_SIZE {
+                return if bytes_read == (RESPONSE_SIZE as usize) {
                         Ok(self.prepare_data_from_buffer())
                 } else {
                     let err_type = ErrorKind::ConnectionAborted;
